This should remove all build warnings on Windows machines.
Also updated the code to pass strings by reference, using .empty() instead of checking for size equivalence to 0, and const correctness.

+++ src/rcpp_xint_xdev_functions.cpp
@@ -120,7 +120,7 @@ double xdev_add_references(Rcpp::Environment data,
     }
 
     vector<Reference> refs;
-    for (int i = 0; i < reference_names.size(); i++) {
+    for (size_t i = 0; i < reference_names.size(); i++) {
 
         Reference ref(reference_names[i]);
         if (hasVersions) {
